Refactor student ID generation in UserProfile and registration form
