+++ lab/forms.py
@@ -45,17 +45,6 @@ class StudentRegistrationForm(UserCreationForm):
         self.fields['password1'].widget.attrs['placeholder'] = 'Create a password'
         self.fields['password2'].widget.attrs['placeholder'] = 'Confirm your password'
     
-    def generate_student_id(self):
-        """Generate a unique student ID starting with STU"""
-        while True:
-            # Generate random 6-digit number
-            random_number = ''.join(random.choices(string.digits, k=6))
-            student_id = f"STU{random_number}"
-            
-            # Check if this ID already exists
-            if not UserProfile.objects.filter(student_id=student_id).exists():
-                return student_id
-    
     def save(self, commit=True):
         user = super().save(commit=False)
         user.email = self.cleaned_data['email']
@@ -63,11 +52,19 @@ class StudentRegistrationForm(UserCreationForm):
         user.last_name = self.cleaned_data['last_name']
         if commit:
             user.save()
-            UserProfile.objects.create(
+            # Don't create UserProfile here - let the signal handler do it
+            # But if signal fails for some reason, use get_or_create as backup
+            profile, created = UserProfile.objects.get_or_create(
                 user=user,
-                role='student',
-                student_id=self.generate_student_id()
+                defaults={
+                    'role': 'student',
+                    'student_id': None  # Will be generated by signal or below
+                }
             )
+            # If profile already exists but doesn't have student_id, add it
+            if not profile.student_id:
+                profile.student_id = profile.generate_student_id()
+                profile.save()
         return user
 
 class ScenarioForm(forms.ModelForm):
